Clean up driver, only use pvt_geodetic
